feat: :mag: Add slugified urls for posts
Instead of post urls being /posts/<id> where id was an integer, posts now use a slug. This required adding a new slug field to the database as well as modifying all references to the post.id to change them to post.slug.

This could break any reference to a post object. Looks good so far, but could have missed something.