perf: :card_file_box: Modify settings.py to set postgres as the default database
My understanding is that postgres is a more 'production ready' database. It's also a lot easier to check things out in an IDE.

Definitely going to break the app until dependencies are updated...install python connector, add postgress password to config file...oh and install and migrate data to postgres on prod.