style: :mag: Add tiles and meta descriptions to most pages. Replace slice method with truncatechars.
Adding meta descriptions allow search engines to feature a snippet about each page. Titles modify the blue links users click on. The only page I wasn't able to modify was the post-detail page. To add a description to this page, I will need to add a description field to each post.

The slice/truncate operation sometimes breaks posts that have code snippets in them that appear toward the very beginning of the post.